Pomelo MySql and no Views

+++ GarageBet.Api/Controllers/AuthenticationController.cs
@@ -85,7 +85,7 @@ namespace GarageBet.Api.Controllers
             User existingUser = _userRepository.FindByEmail(user.Email);
             if (existingUser != null)
             {
-                return StatusCode(HttpStatusCode.Conflict);
+                return StatusCode((int)HttpStatusCode.Conflict);
             }
 
             try
